backend/requirements.txt
Flask
Flask-SQLAlchemy
flask-cors
python-vlc